requirements.txt
Pillow
civitai-py
fastapi
psycopg[binary,pool]
pydantic
requests
slack_sdk
uvicorn